Iterate 1 full cycle per entity rather than one system at a time to fix order of operations problems

+++ EntityComponentSystemCSharp/Components.cs
@@ -71,9 +71,9 @@ namespace EntityComponentSystemCSharp.Components
 
 	public class Actor : IComponent
 	{
-		public int Speed;
+		public float Speed;
 		public int Gold;
-		public int Energy;
+		public float Energy;
 	}
 
 	public class Glyph : IComponent
